Changes to now be able to connect to the new Database
The way the string is present in the code should be handled diffrently.
Dapper removed from RouteDAO.
Needs to be made a choise regarding if pessimistic or optimistic concurreny control should be done on OrderDAO CreateOrder ( Suggestion. Remove the optimistic part and just keep the isolationlevel at RepeatableRead because of the work that is being done there)